Serialization code for SoundInfo
