+++ src/Media/Audio/AudioPlayer.h
@@ -16,131 +16,7 @@
 #include "Engine/Spells/SpellEnums.h"
 #include "Engine/Objects/ActorEnums.h"
 
-enum SoundID {
-    SOUND_Invalid = 0,
-    SOUND_enter = 6,
-    SOUND_WoodDoorClosing = 7,
-    SOUND_fireBall = 8,
-    SOUND_BoatCreaking = 18,
-    SOUND_ClickMinus = 20,
-    SOUND_ClickMovingSelector = 21,
-    SOUND_ClickPlus = 23,
-    SOUND_ClickSkill = 24,
-    SOUND_error = 27,
-    SOUND_dull_strike = 44,
-    SOUND_metal_vs_metal01h = 45,
-    SOUND_metal_vs_metal03h = 47,
-    SOUND_48 = 48,
-    SOUND_RunBadlands = 49,
-    SOUND_RunCarpet = 50,
-    SOUND_RunCooledLava = 51,
-    SOUND_RunDesert = 52,
-    SOUND_RunDirt = 53,
-    SOUND_RunGrass = 54,
-    SOUND_Run55_94 = 55, // sound like some ground-like surface
-    SOUND_RunGround = 56,
-    SOUND_RunRoad = 57,
-    SOUND_RunSnow = 58,
-    SOUND_Run59_98 = 59, // sound like road walk but duller
-    SOUND_Run60_99 = 60, // sound like indoor stone walk with echo added
-    SOUND_RunSwamp = 61,
-    SOUND_RunWater = 62,
-    SOUND_RunWaterIndoor = 63,
-    SOUND_RunWood = 64,
-    SOUND_Run65_104 = 65, // sound like wood walk but thinner
-    SOUND_SelectingANewCharacter = 66,
-    SOUND_shoot_blaster01 = 67,
-    SOUND_shoot_bow01 = 71,
-    SOUND_StartMainChoice02 = 75,
-    SOUND_swing_with_axe01 = 78,
-    SOUND_swing_with_axe03 = 80,
-    SOUND_swing_with_blunt_weapon01 = 81,
-    SOUND_swing_with_blunt_weapon03 = 83,
-    SOUND_swing_with_sword01 = 84,
-    SOUND_swing_with_sword02 = 85,
-    SOUND_WalkBadlands = 88,
-    SOUND_WalkCarpet = 89,
-    SOUND_WalkCooledLava = 90,
-    SOUND_WalkDesert = 91,
-    SOUND_WalkDirt = 92,
-    SOUND_WalkGrass = 93,
-    SOUND_Walk55_94 = 94, // sound like some ground-like surface
-    SOUND_WalkGround = 95,
-    SOUND_WalkRoad = 96,
-    SOUND_WalkSnow = 97,
-    SOUND_Walk59_98 = 98, // sound like road walk but duller
-    SOUND_Walk60_99 = 99, // sound like indoor stone walk with echo added
-    SOUND_WalkSwamp = 100,
-    SOUND_WalkWater = 101,
-    SOUND_WalkWaterIndoor = 102,
-    SOUND_WalkWood = 103,
-    SOUND_Walk65_104 = 104, // sound like wood walk but thinner
-    SOUND_metal_armor_strike1 = 105,
-    SOUND_metal_armor_strike2 = 106,
-    SOUND_metal_armor_strike3 = 107,
-    SOUND_dull_armor_strike1 = 108,
-    SOUND_dull_armor_strike2 = 109,
-    SOUND_dull_armor_strike3 = 110,
-    SOUND_bricks_down = 120,
-    SOUND_bricks_up = 121,
-    SOUND_damage = 122,
-    SOUND_deal = 123,
-    SOUND_defeat = 124,
-    SOUND_querry_up = 125,
-    SOUND_querry_down = 126,
-    SOUND_shuffle = 127,
-    SOUND_title = 128,
-    SOUND_tower_up = 129,
-    SOUND_typing = 130,
-    SOUND_victory = 131,
-    SOUND_wall_up = 132,
-    SOUND_luteguitar = 133,
-    SOUND_panflute = 134,
-    SOUND_trumpet = 135,
-    SOUND_gold01 = 200,
-    SOUND_heal = 202,
-    SOUND_fizzle = 203,
-    SOUND_TurnPage1 = 204,
-    SOUND_TurnPage2 = 205,
-    SOUND_batlleen = 206,
-    SOUND_batllest = 207,
-    SOUND_openchest0101 = 208,
-    SOUND_spellfail0201 = 209,
-    SOUND_drink = 210,
-    SOUND_eat = 211,
-    SOUND_gong = 215,
-    SOUND_hurp = 217,
-    SOUND_church = 218,
-    SOUND_chimes = 219,
-    SOUND_splash = 220,
-    SOUND_star1 = 221,
-    SOUND_star2 = 222,
-    SOUND_star4 = 224,
-    SOUND_eradicate = 225,
-    SOUND_eleccircle = 226,
-    SOUND_encounter = 227,
-    SOUND_openbook = 230,
-    SOUND_closebook = 231,
-    SOUND_teleport = 232,
-    SOUND_hf445a = 5788,
-    SOUND_Haste = 10040,
-    SOUND_21fly03 = 11090,
-    SOUND_WaterWalk = 12040,
-    SOUND_Stoneskin = 13040,
-    SOUND_Bless = 14010,
-    SOUND_Fate = 14020,
-    SOUND_51heroism03 = 14060,
-    SOUND_94dayofprotection03 = 17070,
-    SOUND_9armageddon01 = 17080,
-    SOUND_Sacrifice2 = 18060,
-    SOUND_quest = 20001,
-};
-
-enum MusicID {
-    MUSIC_Invalid = 0,
-    MUSIC_MainMenu = 14,
-    MUSIC_Credits = 15
-};
+#include "SoundEnums.h"
 
 struct AudioSamplePoolEntry {
     AudioSamplePoolEntry(PAudioSample samplePtr_, SoundID id_, int pid_):samplePtr(samplePtr_), id(id_), pid(pid_) {}
@@ -322,8 +198,6 @@ class AudioPlayer {
     AudioSamplePool _loopingSoundPool;
     PAudioSample _currentWalkingSample;
 };
-
-struct SoundDesc;
 
 class SoundList {
  public:
